refactor: flag generation code
