+++ package.json
@@ -36,7 +36,6 @@
     "prebuild:prep:publish": "echo 'building for publishing'",
     "prebuild:theme": "echo 'start generating styles'",
     "prebuild": "echo 'start cleaning up' && rm -rf dist && echo 'done cleaning up'",
-    "presvgs:react:flags": "ts-node ./svgReact/flags/runGenerateSvgs.ts",
     "proxy": "pnpm proxy:options --source 6007 --target 6006",
     "proxy:options": "NODE_OPTIONS=--dns-result-order=ipv4first local-ssl-proxy",
     "push:tags": "git push --follow-tags",
@@ -44,7 +43,7 @@
     "start": "pnpm storybook build && pnpm storybook",
     "storybook": "storybook dev -p 6006 --no-open",
     "storybook:build": "storybook build",
-    "svgs:react:flags": "ts-node ./svgReact/flags/runCreateComponentsFromSvgs.ts",
+    "svgs:react:flags": "ts-node ./svgReact/flags",
     "svgs:react:icons": "ts-node ./svgReact/icons",
     "template:logic-constant": "ts-node build-utils/css/generateLogicConstTemplate.ts",
     "test:prod": "vitest run",
