+++ src/all.ts
@@ -188,15 +188,20 @@ import './icons/CopyIcon';
 import './icons/CubeDuotoneIcon';
 import './icons/CubeIcon';
 import './icons/custom/AvatarIcon';
-import './icons/custom/products/CheckoutIcon';
-import './icons/custom/products/ClassifyIcon';
-import './icons/custom/products/ClearIcon';
-import './icons/custom/products/DashboardIcon';
-import './icons/custom/products/HelloIcon';
-import './icons/custom/products/LandedCostIcon';
-import './icons/custom/products/RateIcon';
-import './icons/custom/products/RestrictIcon';
-import './icons/custom/products/ScreenIcon';
+import './icons/custom/products/CheckoutColorIcon';
+import './icons/custom/products/ClassifyColorIcon';
+import './icons/custom/products/ClearColorIcon';
+import './icons/custom/products/CounterCompanionColorIcon';
+import './icons/custom/products/DashboardColorIcon';
+import './icons/custom/products/DDPColorIcon';
+import './icons/custom/products/DDUColorIcon';
+import './icons/custom/products/HelloColorIcon';
+import './icons/custom/products/InclusivePricingColorIcon';
+import './icons/custom/products/LandedCostColorIcon';
+import './icons/custom/products/RateColorIcon';
+import './icons/custom/products/RestrictColorIcon';
+import './icons/custom/products/ScreenColorIcon';
+import './icons/custom/products/TaxColorIcon';
 import './icons/custom/theme/ThemeDarkIcon';
 import './icons/custom/theme/ThemeLightIcon';
 import './icons/custom/us-states/AlabamaIcon';
@@ -679,7 +684,6 @@ import './utils/getCountryCodeByName';
 import './utils/getFuzzySearch';
 import './utils/getHashId';
 import './utils/getIsInternalLink';
-import './utils/getProductDetails';
 import './utils/getTestId';
 import './utils/handleFetch';
 import './utils/hooks/action-pivot-table/useHasuraGqlPagination';
