--- svgReact/flags/runCreateComponentsFromSvgs.ts
@@ -1,7 +1,0 @@
-import { createComponentsFromSvgs } from 'svgReact/flags/createComponentsFromSvgs';
-
-const run = async () => {
-  await createComponentsFromSvgs();
-};
-
-run();
