--- svgReact/flags/runGenerateSvgs.ts
@@ -1,7 +1,0 @@
-import { generateSvgs } from 'svgReact/flags/generateSvgs';
-
-const run = async () => {
-  await generateSvgs();
-};
-
-run();
